+++ pkg/services/ngalert/provisioning/contactpoints.go
@@ -1,4 +1,4 @@
-package services
+package provisioning
 
 import (
 	"context"
